requirements.txt
pocketflow>=0.0.1
pyyaml>=6.0
requests>=2.28.0
gitpython>=3.1.0
google-cloud-aiplatform>=1.25.0
google-genai>=1.9.0
python-dotenv>=1.0.0
pathspec>=0.11.0
tiktoken
openai
tree-sitter-python
tree-sitter-languages
neo4j


dotenv
requests
boto3
openai
bs4
crawl4ai
aiofiles
psutil
browser-use
browser-use[memory]
asyncio
pydantic

google-api-python-client
youtube-transcript-api

numpy
scikit-learn>=1.3.0     

python-dotenv>=1.0.0          
aiohttp>=3.8.5                
asyncio-throttle>=1.0.2      